Update README.md to reflect a more nuanced language
This improvement will massively improve the likelihood of a foreign developer misinterpreting your scuffed introductory sentence. The current form of the commencing paragraph is unlikely to be correctly construed by any developers from less educated countries, as it can be quite hard to render the meaning of the sentence in question without the proper understanding of the English language.